Moving subcommands to separate folder.
Adding a shared set of decorators for shared options or arguments.
