+++ grow/commands/subcommands/convert.py
@@ -1,19 +1,22 @@
+"""Subcommand for converting pod files from earlier grow versions."""
+
+import os
+import click
+from grow.commands import shared
 from grow.pods import pods
 from grow.pods import storage
 from grow.conversion import content_locale_split
-import click
-import os
 
 
 @click.command()
-@click.argument('pod_path', default='.')
-@click.option('--type', type=click.Choice(['content_locale_split']))
-def convert(pod_path, type):
+@shared.pod_path_argument
+@click.option('--type', 'convert_type', type=click.Choice(['content_locale_split']))
+def convert(pod_path, convert_type):
     """Converts pod files from an earlier version of Grow."""
     root = os.path.abspath(os.path.join(os.getcwd(), pod_path))
     pod = pods.Pod(root, storage=storage.FileStorage)
 
-    if type == 'content_locale_split':
+    if convert_type == 'content_locale_split':
         content_locale_split.Converter.convert(pod)
     else:
         raise click.UsageError(
