+++ grow/commands/subcommands/stage.py
@@ -1,16 +1,19 @@
+"""Subcommand for staging pod to remote server."""
+
+import os
+import click
+from grow.commands import shared
 from grow.common import utils
 from grow.deployments import stats
 from grow.deployments.destinations import base
 from grow.deployments.destinations import webreview_destination
 from grow.pods import pods
 from grow.pods import storage
-import click
-import os
 
 
 # pylint: disable=too-many-locals
 @click.command()
-@click.argument('pod_path', default='.')
+@shared.pod_path_argument
 @click.option('--preprocess/--no-preprocess', '-p/-np', default=True,
               is_flag=True, help='Whether to run preprocessors.')
 @click.option('--remote',
