+++ grow/commands/subcommands/stats.py
@@ -1,12 +1,15 @@
+"""Subcommand for displaying pod stats."""
+
+import os
+import click
+from grow.commands import shared
 from grow.deployments import stats as stats_lib
 from grow.pods import pods
 from grow.pods import storage
-import click
-import os
 
 
 @click.command()
-@click.argument('pod_path', default='.')
+@shared.pod_path_argument
 @click.option('--full/--no-full', '-f', is_flag=True, default=False,
               help='Whether to show full stats. By default, only '
                    'short stats are displayed. Short stats do not '
@@ -17,8 +20,8 @@ def stats(pod_path, full):
     pod = pods.Pod(root, storage=storage.FileStorage)
     try:
         with pod.profile.timer('grow_stats'):
-            stats = stats_lib.Stats(pod, full=full)
-            click.echo_via_pager('\n\n'.join(stats.to_tables()))
-    except pods.Error as e:
-        raise click.ClickException(str(e))
+            pod_stats = stats_lib.Stats(pod, full=full)
+            click.echo_via_pager('\n\n'.join(pod_stats.to_tables()))
+    except pods.Error as err:
+        raise click.ClickException(str(err))
     return pod
